Tvidano/fix csv lateral control (#3)
Fixes the bug in `lateral_controller.py` that prevents csv logging from
failing. Now the signals being updated in the callbacks are stored in a
method scope before using them to compute. Also improved the loginfo
messages so it is more clear what the node is doing during logging and
during shutdown.

---------

Signed-off-by: tvidano <tvidano@gmail.com>