Fix lateral controller error calculation bug (#7)
A previous commit changed the way that the `AbstractLateralController`
computes the cross-track error and introduced a gain of -1. This works
in the Youla controller but breaks the Stanley controller. The fix is to
revert the error computation change and add a gain of -1 to the Youla
controller only. This solution fixes both the Stanley Controller and
Youla Controller.

Signed-off-by: tvidano <61514829+tvidano@users.noreply.github.com>